added a package for sharing context between edit and non edit mode
