+++ .changeset/thick-jobs-call.md
@@ -1,0 +1,5 @@
+---
+'@tinacms/toolkit': patch
+---
+
+Ensure "src" is used instead of "previewSrc" for images in MDX
