+++ packages/@tinacms/sharedctx/README.md
@@ -1,0 +1,4 @@
+# @tinacms/sharedctx
+
+This package is only used internally and is for sharing context from non edit mode and edit mode. This package should be **very** small as it will end it up the users production bundle.
+
