+++ packages/@tinacms/sharedctx/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@tinacms/sharedctx",
+  "version": "0.0.1",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "bugs": {
+    "url": "https://github.com/tinacms/tinacms/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tinacms/tinacms.git",
+    "directory": "packages/@tinacms/sharedctx"
+  },
+  "license": "Apache-2.0",
+  "scripts": {
+    "types": "yarn tsc",
+    "build": "echo \"Run `yarn build` from the root of the repository instead\""
+  },
+  "devDependencies": {
+    "@tinacms/scripts": "workspace:*",
+    "react": "17.0.2",
+    "react-dom": "17.0.2"
+  },
+  "peerDependencies": {
+    "react": ">=16.14",
+    "react-dom": ">=16.14",
+    "react-is": ">=16.14.0"
+  }
+}
