+++ packages/@tinacms/toolkit/src/packages/fields/plugins/MdxFieldPlugin/plate/toolbar/index.tsx
@@ -136,7 +136,7 @@ export const ToolbarButtons = ({ name, templates }) => {
         [
           {
             type: 'img',
-            url: selectedMedia.previewSrc,
+            url: selectedMedia.src || selectedMedia.previewSrc,
             alt: '',
             caption: '',
             children: [
